requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyodbc==4.0.39
python-dotenv==1.0.0
pydantic==2.4.2
